backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pyngrok
